Client: Rewrote onCommand in an extensible way - it's about 100x as awesome now and uses reflection. There is still some duplication between cmdHandlers and the Protocol class, but it's not bad and I can't think of something better.
Added declerations and initializations of some of the data structures we talked about for caching. Not sure how the pending requests queue might work yet in regards to timeouts - when does it get checked - and what we actually put in it.

Changed printError methods to use stringBuilders beacuse addings lots of strings sucks.

Added some TODO comments for project 2.

Reorganized a few other things and added some comments.

I think we should factor all of the FS operations out of the Client and into their own class to make everything a little easier to look at.